Implement a new feature, Link Selection Policy. Fixed mutliple bugs. Re-arrange class names.

+++ src/org/cloudbus/cloudsim/sdn/AggregationSwitch.java
@@ -17,7 +17,7 @@ package org.cloudbus.cloudsim.sdn;
  */
 public class AggregationSwitch extends Switch {
 
-	public AggregationSwitch(String name,int bw, long iops, int upports, int downports, NetworkOperatingSystem nos) {
+	public AggregationSwitch(String name,long bw, long iops, int upports, int downports, NetworkOperatingSystem nos) {
 		super(name, bw, iops, upports, downports, nos);
 	}
