+++ src/org/cloudbus/cloudsim/sdn/CoreSwitch.java
@@ -17,7 +17,7 @@ package org.cloudbus.cloudsim.sdn;
  */
 public class CoreSwitch extends Switch {
 
-	public CoreSwitch(String name,int bw, long iops, int upports, int downports, NetworkOperatingSystem nos) {
+	public CoreSwitch(String name,long bw, long iops, int upports, int downports, NetworkOperatingSystem nos) {
 		super(name, bw, iops, upports, downports, nos);
 		//if (upports>0) throw new IllegalArgumentException("Core switches cannot have uplinks.");
 	}
