+++ src/org/cloudbus/cloudsim/sdn/EdgeSwitch.java
@@ -17,7 +17,7 @@ package org.cloudbus.cloudsim.sdn;
  */
 public class EdgeSwitch extends Switch {
 
-	public EdgeSwitch(String name,int bw, long iops, int upports, int downports, NetworkOperatingSystem nos) {
+	public EdgeSwitch(String name,long bw, long iops, int upports, int downports, NetworkOperatingSystem nos) {
 		super(name, bw, iops, upports, downports, nos);
 	}
